+++ app/src/main/java/woowacourse/shopping/data/repository/okhttp/CartRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.repository
+package woowacourse.shopping.data.repository.okhttp
 
 import woowacourse.shopping.data.mapper.toData
 import woowacourse.shopping.data.mapper.toDomain
