--- app/src/main/java/woowacourse/shopping/util/RetrofitUtil.kt
@@ -1,16 +1,0 @@
-package woowacourse.shopping.util
-
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-
-object RetrofitUtil {
-    val okHttpClient: OkHttpClient by lazy {
-        OkHttpClient.Builder()
-            .addInterceptor(
-                HttpLoggingInterceptor().apply {
-                    level = HttpLoggingInterceptor.Level.BODY
-                },
-            )
-            .build()
-    }
-}
